chore: reverts ff for sidebar from main
