+++ src/components/AppLayout/menuStructure.ts
@@ -22,14 +22,14 @@ import { pageListPath } from "@dashboard/pages/urls";
 import { SidebarMenuItem } from "@saleor/macaw-ui";
 import { IntlShape } from "react-intl";
 
-import { appsListPath } from "../../../apps/urls";
-import { categoryListUrl } from "../../../categories/urls";
-import { collectionListUrl } from "../../../collections/urls";
-import { customerListUrl } from "../../../customers/urls";
-import { saleListUrl, voucherListUrl } from "../../../discounts/urls";
-import { orderDraftListUrl, orderListUrl } from "../../../orders/urls";
-import { productListUrl } from "../../../products/urls";
-import { languageListUrl } from "../../../translations/urls";
+import { appsListPath } from "../../apps/urls";
+import { categoryListUrl } from "../../categories/urls";
+import { collectionListUrl } from "../../collections/urls";
+import { customerListUrl } from "../../customers/urls";
+import { saleListUrl, voucherListUrl } from "../../discounts/urls";
+import { orderDraftListUrl, orderListUrl } from "../../orders/urls";
+import { productListUrl } from "../../products/urls";
+import { languageListUrl } from "../../translations/urls";
 import { getMenuItemExtension, mapToExtensionsItems } from "./utils";
 
 export interface FilterableMenuItem extends Omit<SidebarMenuItem, "children"> {
