--- src/components/Sidebar/Drawer.tsx
@@ -1,10 +1,0 @@
-import { useFlags } from "@dashboard/hooks/useFlags";
-import React from "react";
-
-import { LegacyDrawer } from "./legacy";
-
-export const SidebarDrawer = () => {
-  const { enableNewSidebar } = useFlags(["enableNewSidebar"]);
-
-  return enableNewSidebar.enabled ? null : <LegacyDrawer />;
-};
