--- src/components/Sidebar/NewSidebar.tsx
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const NewSidebar = () => <div>Work in progress</div>;
