--- src/components/Sidebar/index.ts
@@ -1,2 +1,0 @@
-export * from "./Drawer";
-export * from "./Sidebar";
