--- src/components/Sidebar/legacy/index.ts
@@ -1,2 +1,0 @@
-export * from "./LegacyDrawer";
-export * from "./LegacySidebar";
